# swagger.yaml
openapi: 3.0.0
info:
  title: Webser2 API
  description: Web servisi analitik ve izleme API'si
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your API key as the Bearer token
  schemas:
...
    JobStatus:
      type: object
      properties:
        jobId:
          type: string
          example: "123"
        status:
          type: string
          enum: [queued, processing, completed, failed]
          example: "completed"
        progress:
          type: integer
          minimum: 0
          maximum: 100
          example: 100
        startTime:
          type: string
          format: date-time
        finishTime:
          type: string
          format: date-time
        error:
          type: string
        downloadUrl:
          type: string
          example: "/api/job/result/123"
security:
  - BearerAuth: []
paths:
...
  /api/job/status/{jobId}:
    get:
      summary: Get the current status of a query job
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
          description: Job identifier
      responses:
        '200':
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/job/result/{jobId}:
    get:
      summary: Download job results
      description: Download the query results as a ZIP file containing JSON
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
          description: Job identifier
      responses:
        '200':
          description: Results downloaded successfully
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...